Add files via upload
The “Object classification through vocal assistance for visually impaired” project aims to provide awareness of the surrounding environment to a blind person. Visual impairment severely limits the quality of life of the person. According to the World Health Organization (WHO), there are 36 million blind people worldwide , of which India has about 8 million . According to National Health Services (NHS)in England , there are almost 2 million people living with sight loss in UK, with around 360,000 registered as blind or partially sighted. A technology-assisted device capable of providing accurate awareness assistance to the blind person can support these people in daily lives, for example, they can use the technology-assisted awareness device to freely move around without requiring a family member's help. This will significantly improve their quality of life as they will need a reduced amount of support from other people and can become more independent. 
The image captured by the portable scanner is accessed through a mobile application. The object in the image is recognized and given out as Audio. For example, when the scanner captures anything (such as chair, table, apple, or any object that he/she may encounter in daily life), the application in the smartphone will voice out (speak) the recognized object’s name. 

n this project, we will implement awareness assistance such as object and facial recognition that will cater to assist the huge population of the blind people. According to WHO, there are 36 million blinds worldwide, of which India is home to about 8 million. According to the NHS in England, there are almost 2 million people living with sight loss in the UK, with around 360,000 registered as blind or partially sighted. “Lack of any sense” is a quoted terminology for having an impartial life. This project aims to develop a cost-effective technology assisted device that increases the quality of life of the blind by making them independent and possibly empowering them. Presently, at the end of funding, this device will produce a spark of artificial regeneration (mental model) of sight via facial and object detection. Though visualizing is not solved, knowing what or who is present in front of the blind person would sound like the beginning of a promising lifestyle for them. Being aware of one’s surroundings would help the blind person reorganize their life. With several enhancements as discussed in the project details, this project will sustain for a long term. 

The project is aimed at satisfying the keen demand of the huge blind population for better “seeing” the world. This awareness unit will consist of a camera  to capture an image which is then accessed through a mobile application.  The image is further recognized and classified to get the audio output. We used the YOLOv3 algorithm for object detection, which is a state-of-the-art deep learning model that can detect objects in real-time. YOLOv3 works by dividing the input image into a grid of cells and predicting the objectness score and bounding boxes for each cell.